Initial commit, with minimal functionality.
When run from the AssettoServer directory, automatically loads the
entries file, looks for any that meet specific criteria and adds a key
used by AssettoServer to indicate that they can be used as traffic.

The current criteria are that the car's model includes "traffic" (e.g.
it's from the traffic pack), its team contains "traffic", or its team
contains "reserve".